+++ summoner-cli/examples/full-batteries/benchmark/Main.hs
@@ -1,0 +1,7 @@
+module Main (main) where
+
+import Gauge.Main
+
+
+main :: IO ()
+main = defaultMain [bench "const" (whnf const ())]