+++ summoner-cli/examples/full-batteries/src/FullBatteries.hs
@@ -1,0 +1,15 @@
+{- |
+Copyright: (c) 2020 Kowainik
+SPDX-License-Identifier: MIT
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+Full-featured test project
+-}
+
+module FullBatteries
+       ( someFunc
+       ) where
+
+
+someFunc :: IO ()
+someFunc = putStrLn ("someFunc" :: String)