+++ summoner-cli/examples/full-batteries/test/Spec.hs
@@ -1,0 +1,5 @@
+module Main (main) where
+
+
+main :: IO ()
+main = putStrLn ("Test suite is not implemented" :: String)