+++ summoner-cli/examples/stack-full/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-`fullProject` uses [PVP Versioning][1].
+`stack-full` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
 ## 0.0.0.0
@@ -8,4 +8,4 @@ The changelog is available [on GitHub][2].
 * Initially created.
 
 [1]: https://pvp.haskell.org
-[2]: https://github.com/kowainik/fullProject/releases
+[2]: https://github.com/kowainik/stack-full/releases